refactor: HttpMethod 파일 생성 및 HttpInfo을 OpenMarketAPI로 타입명 변경
